Add first draft of Python-JS bridge using zmq
