+++ backend_ts/package.json
@@ -18,6 +18,7 @@
   "dependencies": {
     "@nxg-org/mineflayer-util-plugin": "^1.8.3",
     "mineflayer": "^4.26.0",
-    "mineflayer-pathfinder": "^2.4.5"
+    "mineflayer-pathfinder": "^2.4.5",
+    "zeromq": "^6.3.0"
   }
 }
