+++ backend_ts/test/cli.ts
@@ -1,16 +1,16 @@
 import { createPlugin } from "../src";
 import { Bot, createBot } from "mineflayer";
 import readline from "readline";
-import { handleCommand, handleFunctionCall } from "./common";
+import { handleCommand } from "./common";
 
 const bot = createBot({
-  username: "SemanticSteve",
+  username: "test",
 });
 
 bot.once("spawn", () => {
   const semanticSteve = createPlugin({ immediateRadius: 5, nearbyRadius: 128 });
   bot.loadPlugin(semanticSteve);
-  console.log("Bot spawned and ready!");
+  console.log("Bot spawned and ready for interactive CLI commands.");
   startCLI();
 });
 
@@ -21,21 +21,25 @@ const rl = readline.createInterface({
 });
 
 
-function startCLI() {  
-  console.log("STATE:", bot.semanticWorld.toString())
-  rl.setPrompt("Please enter your function call (JS syntax): ");
+function generatePrompt(bot: Bot) {
+  return `Semantic Steve 1.0.0!
+  World State:
+  ${bot.semanticWorld.toString()}
+
+  [${bot.username}] > `;
+}
+
+
+function startCLI() {
+  rl.setPrompt(generatePrompt(bot));
   rl.prompt();
 
-  let result: any;
-  rl.on("line", async (line) => {  // E.g., pathfindToCoordinate([-585.5, 68, 54.5])
-    result = await handleFunctionCall(bot, line);
-    console.log("RESULT:", result);
-    console.log("STATE:", bot.semanticWorld.toString())
-    rl.setPrompt("Please enter your function call (JS syntax): ");
+  rl.on("line", async (line) => {
+    await handleCommand(bot, line);
+    rl.setPrompt(generatePrompt(bot));
     rl.prompt();
   });
 }
-
 
 bot.on("chat", (username, message) => {
   if (username === bot.username) return;
